requirements.txt
asgiref==3.2.5
Django==3.0.4
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
Keras==2.1.3
tensorflow==1.14.0
spacy
opencv-python
# python3 -m spacy download en_core_web_sm
# python3 -m spacy download en_vectors_web_lg
